Support for a Configuration Stream Handler
I noticed in the code for all the stream managed configuration sources
(Ini, Json, XML) a lot of duplicated code.  So i abstracted that down to
a BaseStreamConfigurationSource for them to inherit from.  This helps
keep the code a little cleaner, provides a foundation for creating other
Stream based configuration sources in a common way.  Also with the new
IConfigurationStreamHandler interface, any number of sources for the
files could be used (File (default), Azure Blob, Document DB)..  My next
goal would be to get a blob and document DB version of
IConfigurationStreamHandler..  And then a set of extension methods to
streamline adding these sources to configuration.